refactor: Drop withDiff(), rewrite TranslationDiff
