+++ translate/src/core/diff/components/TranslationDiff.test.js
@@ -1,21 +1,21 @@
 import React from 'react';
-import { shallow } from 'enzyme';
+import { mount } from 'enzyme';
 
-import TranslationDiff from './TranslationDiff';
+import { TranslationDiff } from './TranslationDiff';
 
 describe('<TranslationDiff>', () => {
   it('returns the correct diff for provided strings', () => {
-    const wrapper = shallow(
+    const wrapper = mount(
       <TranslationDiff base={'abcdef'} target={'cdefgh'} />,
     );
 
     expect(wrapper.find('ins')).toHaveLength(1);
     expect(wrapper.find('del')).toHaveLength(1);
-    expect(wrapper.at(1).text()).toEqual('cdef');
+    expect(wrapper.childAt(1).text()).toEqual('cdef');
   });
 
   it('returns the same string if provided strings are equal', () => {
-    const wrapper = shallow(
+    const wrapper = mount(
       <TranslationDiff base={'abcdef'} target={'abcdef'} />,
     );
