+++ translate/src/core/diff/index.ts
@@ -1,2 +1,1 @@
-export { default as TranslationDiff } from './components/TranslationDiff';
-export { default as withDiff } from './withDiff';
+export { TranslationDiff } from './components/TranslationDiff';
